feature(timestamp): round order amounts values to two decimals

+++ src/main/java/com/proyect/masterdata/dto/OrderDTO.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -31,10 +32,10 @@ public class OrderDTO {
     private String reference;
     private List<String> paymentReceipts;
     private List<String> courierPictures;
-    private Double saleAmount;
-    private Double deliveryAmount;
-    private Double advancedPayment;
-    private Double duePayment;
+    private BigDecimal saleAmount;
+    private BigDecimal deliveryAmount;
+    private BigDecimal advancedPayment;
+    private BigDecimal duePayment;
     private Date registrationDate;
     private Date updateDate;
     private String deliveryAddress;
